+++ lib/karafka/web.rb
@@ -6,7 +6,6 @@ require "etc"
 require "open3"
 require "zlib"
 require "securerandom"
-require "cgi"
 require "uri"
 
 module Karafka